Create component factory
